+++ src-main/utils/content-formats.js
@@ -1,9 +1,9 @@
-const SUPPORTED_CONTENT_EXTENSIONS = ['md','markdown','html'];
+const SUPPORTED_CONTENT_EXTENSIONS = ['md','markdown','html', 'qmd'];
 
 module.exports.SUPPORTED_CONTENT_EXTENSIONS = SUPPORTED_CONTENT_EXTENSIONS;
 
 module.exports.allValidContentFilesExt = function(filePath ){
-  return filePath && (filePath.endsWith('.md') || filePath.endsWith('.markdown'));
+  return filePath && (filePath.endsWith('.md') || filePath.endsWith('.markdown') || filePath.endsWith('.markdown'));
 }
 
 module.exports.isContentFile = function(filePath ){